GoldPlus를 float로 — 처치 골드 공식이 int 절삭으로 조용히 과소해지고 있었다
원작 Gold_Math × (2 + Gold_Plus)의 Gold_Plus는 real이고 증가량이 전부 소수
(+0.60/+0.30/+0.20/+0.20)다. int였던 지금 구조로는 넷을 다 모아도 1.30인데
int 캐스팅에서 1로 버려져 배수가 10% 과소 계상될 수 있었다. 버림 위치도
잘못돼 있었다 — 원작은 R2I(곱한 뒤 한 번만 버림)인데 지금은 곱하기 전에
(int)로 버렸다. Gold_Math 자체는 원작에서도 정수 나눗셈이라 그대로 뒀다.

GoldPlus를 채우는 조건 넷(나미 A04X 레벨1·버기 포에버+누적10·나미
이터널·아이템 I00Z)은 리서치담당이 찾아뒀지만 포에버·이터널·누적 클리어
개념이 우리에 아직 없어 배선하지 않았다(PM 지시) — 주석에 남겨 다음에
조건만 채우면 되게 해뒀다. 지금은 GoldPlus가 항상 0이라 이 커밋 자체는
동작 회귀가 없다.

compile_check.sh 통과.

+++ Assets/Scripts/Units/GoldWallet.cs
@@ -7,10 +7,17 @@ public class GoldWallet : MonoBehaviour
     public int Gold { get; private set; }
 
     // 항법(원작 route 선택) 보너스 — 처치 골드 공식 Gold_Math × (2 + Gold_Plus)의 두 번째 항이다.
-    // 리서치담당이 조건(나미 전설 조합 +0.6, 영원함 나미/버기 택1 +0.2~0.3, 수배서 아이템 +0.2)을
-    // 찾아뒀지만 어디서 어떻게 걸리는지는 아직 안 정해져서 항상 0이다(PM 지시, 2026-09-04) —
-    // 정해지면 이 필드를 채우는 코드만 새로 붙이면 된다. 공식(RewardDistributor) 쪽은 이미 준비됨.
-    public int GoldPlus { get; private set; }
+    // 원작은 real이고 증가량이 전부 소수라 float이어야 한다(2026-09-06 PM 확인, int였을 때
+    // 넷을 다 모아도 1.30 → int면 1로 버려져 최종 배수가 10% 과소해졌다 — 지금은 아래 넷이
+    // 전부 미배선이라 항상 0이므로 이 교정 자체는 회귀가 없다).
+    // 조건 넷(리서치담당 확보, 2026-09-04/06):
+    //   +0.60  A04X 레벨==1 AND Nami_legend_Boolean==false
+    //   +0.30  버기 포에버 AND 누적 클리어 ≥ 10
+    //   +0.20  나미 이터널
+    //   +0.20  포인트값 22 유닛이 아이템 I00Z 획득
+    // 포에버·이터널·누적 클리어 개념이 우리에 없어 아직 배선 안 함(PM 지시) — 정해지면
+    // 이 필드를 채우는 코드만 새로 붙이면 된다. 공식(RewardDistributor) 쪽은 이미 준비됨.
+    public float GoldPlus { get; private set; }
 
     public event System.Action<int> OnGoldChanged;
